gateway: move validation-error classifier to api layer
Per @arner's review: core no longer imports api/rpcerr. Sentinels
(ErrUnprotectedTx, ErrNonceLookup) move to gateway/domain; the
core->api translation lives in gateway/api/rpcerrors.go and runs
inside SendRawTransaction.
