+++ gateway/core/validate.go
@@ -8,7 +8,6 @@ package core
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"math"
 	"math/big"
@@ -18,16 +17,12 @@ import (
 	"github.com/ethereum/go-ethereum/core/txpool"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/params"
+	"github.com/hyperledger/fabric-x-evm/gateway/domain"
 )
 
 type stateReader interface {
 	NonceAt(ctx context.Context, account common.Address, blockNumber *big.Int) (uint64, error)
 }
-
-var (
-	errUnprotectedTx = errors.New("only replay-protected (EIP-155) transactions allowed over RPC")
-	errNonceLookup   = errors.New("look up nonce")
-)
 
 // txMaxSize redeclares the unexported core/txpool/legacypool constant (4 * 32 KiB).
 const txMaxSize = 4 * 32 * 1024
@@ -50,7 +45,7 @@ func ValidateTx(
 	// Geth rejects this in internal/ethapi.SubmitTransaction, above the txpool —
 	// the txpool's signer recovery accepts Frontier-style signatures.
 	if !tx.Protected() {
-		return errUnprotectedTx
+		return domain.ErrUnprotectedTx
 	}
 
 	head := &types.Header{
@@ -77,7 +72,7 @@ func ValidateTx(
 
 	nonce, err := state.NonceAt(ctx, from, nil)
 	if err != nil {
-		return fmt.Errorf("%w: %w", errNonceLookup, err)
+		return fmt.Errorf("%w: %w", domain.ErrNonceLookup, err)
 	}
 	if nonce > tx.Nonce() {
 		return fmt.Errorf("%w: next nonce %d, tx nonce %d", ethcore.ErrNonceTooLow, nonce, tx.Nonce())
